+++ lib/amocrm_client/stor_adapters/redis.rb
@@ -2,15 +2,18 @@
 
 module AmocrmClient
   module StorAdapters
-    module Redis
-      extend self
+    class Redis
+      def initialize(config)
+        @config = config
+        @redis_connection = AmocrmClient.redis_connection
+      end
 
       def update(params)
         tokens_set(params)
       end
 
       def find
-        tokens = AmocrmClient.redis_connection.hgetall(redis_key)
+        tokens = redis_connection.hgetall(redis_key)
         tokens.empty? ? nil : tokens
       end
 
@@ -20,12 +23,14 @@ module AmocrmClient
 
       private
 
+      attr_reader :config, :redis_connection
+
       def redis_key
-        @redis_key ||= AmocrmClient.config.stor_adapter.dig('redis', 'key')
+        @redis_key ||= config.stor_adapter.dig('redis', 'key')
       end
 
       def tokens_set(params)
-        AmocrmClient.redis_connection.mapped_hmset(redis_key, params)
+        redis_connection.mapped_hmset(redis_key, params)
         params
       end
     end