+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# [1.0.0] - 2020-09-21
+# [1.2.0] - 2021-03-25
+
+### Added
+  - Supporting several accounts
+
+# [1.0.1] - 2020-09-03
+
+### Added
+  - Redlock for refreshing tokens
+
+# [1.0.0] - 2020-09-01
 
 ### Added
   - First release